OpeningWindow and the VerificationPage if finished. issues with the files has been solved.
