+++ app/src/main/java/com/example/microbankingsystem/MainActivity.java
@@ -5,6 +5,7 @@ import android.os.Bundle;
 
 import com.example.microbankingsystem.ui.OpeningWindow;
 import com.example.microbankingsystem.ui.OptionsFragment;
+import com.example.microbankingsystem.ui.VerificationPage;
 import com.google.android.material.snackbar.Snackbar;
 
 import androidx.appcompat.app.AppCompatActivity;
@@ -28,8 +29,6 @@ public class MainActivity extends AppCompatActivity {
     private ActivityMainBinding binding;
 
     Button btn_test;
-    Button btn_transaction_normal;
-    Button btn_transaction_critical;
 
       @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -54,8 +53,7 @@ public class MainActivity extends AppCompatActivity {
 
         // initializing of buttons
         btn_test = findViewById(R.id.button_test);
-        btn_transaction_normal = findViewById(R.id.btn_transaction_normal);
-        btn_transaction_critical = findViewById(R.id.btn_transaction_critical);
+
 
         // calling setOnClickListner on initialized buttons
         btn_test.setOnClickListener(new View.OnClickListener() {
@@ -64,20 +62,6 @@ public class MainActivity extends AppCompatActivity {
                 testact();
             }
         });
-
-        btn_transaction_normal.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                openOpeningWindow();
-            }
-        });
-
-          btn_transaction_critical.setOnClickListener(new View.OnClickListener() {
-              @Override
-              public void onClick(View view) {
-                  openOpeningWindow();
-              }
-          });
 
     }
 
@@ -111,12 +95,8 @@ public class MainActivity extends AppCompatActivity {
     }
 
     public void testact(){
-        Intent test = new Intent(this, OptionsFragment.class);
+        Intent test = new Intent(this, OpeningWindow.class);
         startActivity(test);
     }
 
-    private void openOpeningWindow(){
-        Intent intent = new Intent(this, OpeningWindow.class);
-        startActivity(intent);
-    }
 }
