+++ app/src/main/java/com/example/microbankingsystem/ui/OpeningWindow.java
@@ -12,9 +12,37 @@ import com.example.microbankingsystem.R;
 
 public class OpeningWindow extends AppCompatActivity {
 
+    Button btn_transaction_normal;
+    Button btn_transaction_critical;
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_opening_window);
+
+        btn_transaction_normal = findViewById(R.id.btn_transaction_normal);
+        btn_transaction_critical = findViewById(R.id.btn_transaction_critical);
+
+
+        btn_transaction_normal.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                openVerificationWindow();
+            }
+        });
+
+        btn_transaction_critical.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                openVerificationWindow();
+            }
+        });
+
     }
+
+    public void openVerificationWindow(){
+        Intent intent = new Intent(this, VerificationPage.class);
+        startActivity(intent);
+    }
+
 }
